Fixed Project Deserializer Project Deserializer would previously throw exceptions trying to deserialize a project, because gson does not support nested json's, and the entirety of each team member's serialization is embedded within the project serialization.
To combat this, we used two helper methods to isolate the owner and team serializations from within the JSON project. These methods use the java String split method to split the JSON string into substrings.

This does have the potential to expose a previously unreachable, but potentially more destructive bug - the duplication of users. We have previously encountered issues of duplication of objects when an objects is included as a member of another object.  I suspect this could occur if we could actually create projects, but that had previously been made impossible by the non-functionality of the ProjectDeserializer.

Signed off by: Alec Thompson, Batyr Nurbekov