Fixed Project Deserializer Project Deserializer was experiencing a UnsupportedOperationException when it attempted to deserialize a project with users included.  To fix this, we used the Java String operation split() to break out the owner User deserialization, and the JSONarray of team members.  To do this we also had to implement a fromJSONarray method for Users, that deserializes a JSONarray string.
We also added a test that uses a real project string, gained from printing the output of one of the serialization tests. This is more realistic than the previous PD tests because it actually includes the owner and team member fields, the first of which always exists. The previous tests only tested the project name and id fields.

Committed by: ajthompson, bnurbekov